fixed some errors occured during commiting privous commit
